feat!: apply changes from next-major comments in code
- update comments to clear out handled TODO items
- remove legacy handling for json/css in theming config
- remove subcomponent handling in story files
- update component type usages (do not use TS interfaces)
- move utility class into Breadcrumbs component
- remove fallback handling of fieldNote in InputField component
- cleanup custom / lingering u-* utility class
- cleanup legacy handling from Tag component
- remove unused CSS mixins
