+++ plop-templates/Component/Component.tsx.hbs
@@ -17,7 +17,7 @@ export type {{pascalCase name}}Props = {
  *
  * `import {{inBraces (pascalCase name)}} from "@chanzuckerberg/eds";`
  *
- * TODO: update this comment with a description of the component.
+ * HERE: update this comment with a description of the component.
  */
 export const {{pascalCase name}} = ({
   className,
