+++ src/components/Accordion/Accordion.module.css
@@ -3,9 +3,6 @@
 /*------------------------------------*\
     # ACCORDION
 \*------------------------------------*/
-/**
- * TODO: Icon inherits color from the surrounding text, but should use the matching -icon- tokens from below
- */
 
 /**
  * Accordion Panel that expands and collapses.
