+++ src/components/AppNotification/AppNotification.module.css
@@ -34,15 +34,11 @@
     margin-bottom: 0.5rem;
 }
 
-/* .app-notification__sub-title {
-} */
-
 .app-notification__actions {
     margin-top: 1.5rem;
 }
 
 .app-notification__content {
-    /* TODO: should this magic number exist as some token? */
     max-width: 1320px;
     display: flex;
     gap: 0.5rem;
