+++ src/components/FieldNote/FieldNote.module.css
@@ -3,9 +3,6 @@
 /*------------------------------------*\
     # FIELD NOTE
 \*------------------------------------*/
-/**
- * TODO: Icon inherits color from the surrounding text, but should use the matching -icon- tokens from below
- */
 
 /**
  * Fieldnote
