+++ src/components/InlineNotification/InlineNotification.module.css
@@ -3,9 +3,6 @@
 /*------------------------------------*\
     # INLINE NOTIFICATION
 \*------------------------------------*/
-/**
- * TODO: Icon inherits color from the surrounding text, but should use the matching -icon- tokens from below
- */
 
 /**
  * A messaging element that is used inline.
