+++ src/components/InputField/InputField.module.css
@@ -1,13 +1,9 @@
 /*------------------------------------*\
     # INPUT FIELD
 \*------------------------------------*/
-/**
- * TODO: Icon inherits color from the surrounding text, but should use the matching -icon- tokens from below
- */
 
 /**
  * Wraps the Label and the optional/required hint.
- * TODO: match/share the overline styles between Select and InputField
  */
 .input-field__overline {
   display: flex;
