+++ src/components/PageNotification/PageNotification.module.css
@@ -3,9 +3,6 @@
 /*------------------------------------*\
     # PAGE NOTIFICATION
 \*------------------------------------*/
-/**
- * TODO: Icon inherits color from the surrounding text, but should use the matching -icon- tokens from below
- */
 
 /**
  * Message of information, success, caution, or warning to the user.
