+++ src/components/PopoverListItem/PopoverListItem.module.css
@@ -1,9 +1,6 @@
 /*------------------------------------*\
     # POPOVER LIST ITEM
 \*------------------------------------*/
-/**
- * TODO: Icon inherits color from the surrounding text, but should use the matching -icon- tokens from below
- */
 
 /**
  * PopoverListItem
