+++ src/components/Select/Select.module.css
@@ -3,9 +3,6 @@
 /*------------------------------------*\
     # SELECT
 \*------------------------------------*/
-/**
- * TODO: Icon inherits color from the surrounding text, but should use the matching -icon- tokens from below
- */
 
 /**
  * Select field used to select one option from a list of options.
