+++ src/components/Tag/Tag.module.css
@@ -25,7 +25,6 @@
 
 /**
  * Spacing between icon and label text, if icon present
- * TODO: perhaps use gap?
  */
 .tag> :not(:last-child) {
   margin-right: calc(var(--eds-spacing-size-half) / 16 * 1rem);
@@ -40,40 +39,6 @@
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
-}
-
-/**
- * Color variants.
- * TODO(next-major): remove these status colors and delete tokens
- */
-:where(.tag--neutral) {
-  --tag-primary-color: var(--eds-theme-color-text-neutral-default);
-  --tag-secondary-color: var(--eds-theme-color-background-neutral-subtle);
-  --tag-outline-color: var(--eds-theme-color-border-neutral-default);
-}
-
-:where(.tag--error) {
-  --tag-primary-color: var(--eds-theme-color-text-utility-critical);
-  --tag-secondary-color: var(--eds-theme-color-background-utility-error);
-  --tag-outline-color: var(--eds-theme-color-border-utility-error-default);
-}
-
-:where(.tag--success) {
-  --tag-primary-color: var(--eds-theme-color-text-utility-favorable);
-  --tag-secondary-color: var(--eds-theme-color-background-utility-success);
-  --tag-outline-color: var(--eds-theme-color-border-utility-success-default);
-}
-
-:where(.tag--warning) {
-  --tag-primary-color: var(--eds-theme-color-text-utility-warning);
-  --tag-secondary-color: var(--eds-theme-color-background-utility-warning);
-  --tag-outline-color: var(--eds-theme-color-border-utility-warning-default);
-}
-
-:where(.tag--brand) {
-  --tag-primary-color: var(--eds-theme-color-text-brand-default);
-  --tag-secondary-color: var(--eds-theme-color-background-brand-primary-default);
-  --tag-outline-color: var(--eds-theme-color-border-brand-primary-default);
 }
 
 /**
@@ -97,12 +62,4 @@
 :where(.tag--status-informational) {
   --tag-primary-color: var(--eds-theme-color-text-utility-informational);
   --tag-secondary-color: var(--eds-theme-color-background-utility-informational-low-emphasis);
-}
-
-/**
- * Adds outline to the tag.
- * TODO(next-major): remove handling for outline in 2.0 component
- */
-.tag--outline {
-  border-color: var(--tag-outline-color);
 }
