+++ src/components/TextareaField/TextareaField.module.css
@@ -3,9 +3,6 @@
 /*------------------------------------*\
     # TEXTAREA FIELD
 \*------------------------------------*/
-/**
- * TODO: Icon inherits color from the surrounding text, but should use the matching -icon- tokens from below
- */
 
 /**
  * Default input styles
