+++ src/components/ToastNotification/ToastNotification.module.css
@@ -3,9 +3,6 @@
 /*------------------------------------*\
     # TOAST
 \*------------------------------------*/
-/**
- * TODO: Icon inherits color from the surrounding text, but should use the matching -icon- tokens from below
- */
 
 .toast {
   width: 100%;
