+++ src/design-tokens/mixins.css
@@ -1,33 +1,11 @@
 /**
  * Legacy Mixins
- * TODO: remove once not in use
+ * TODO(next-major): remove mixin once not in use (after 1.x components have been migrated)
  */
 @define-mixin focus {
   outline: calc(var(--eds-theme-box-focus-ring-outline-width) * 1px) solid
     var(--eds-theme-color-focus-ring);
   outline-offset: calc(var(--eds-theme-box-focus-ring-outline-offset) * 1px);
-}
-
-/* TODO(next-major): remove b/c this is no longer in use */
-@define-mixin focusInverted {
-  @mixin focus;
-  outline-color: var(--eds-theme-color-focus-ring-inverted);
-}
-
-/* TODO(next-major): remove b/c this is no longer in use */
-@define-mixin placeholderStyles {
-  color: var(--eds-theme-color-text-disabled);
-}
-
-/* TODO(next-major): remove this after the one usage is removed */
-@define-mixin visuallyHide {
-  position: absolute !important;
-  overflow: hidden !important;
-  width: 1px !important;
-  height: 1px !important;
-  padding: 0 !important;
-  border: 0 !important;
-  clip: rect(1px, 1px, 1px, 1px) !important;
 }
 
 /**
