+++ src/index.ts
@@ -11,7 +11,6 @@ export { default as Hr } from './components/Hr';
 export { default as Label } from './components/Label';
 export { default as Skeleton } from './components/Skeleton';
 export { default as Table } from './components/Table';
-export { default as Tag } from './components/Tag';
 export { default as Text } from './components/Text';
 export { default as Toggle } from './components/Toggle';
 
@@ -46,6 +45,7 @@ export { default as Radio } from './components/Radio';
 export { default as Select } from './components/Select';
 export { default as SelectionChip } from './components/SelectionChip';
 export { default as TabGroup } from './components/TabGroup';
+export { default as Tag } from './components/Tag';
 export { default as TextareaField } from './components/TextareaField';
 export { default as ToastNotification } from './components/ToastNotification';
 export { default as Tooltip } from './components/Tooltip';
@@ -65,10 +65,6 @@ export type {
   DataTableWithStatus,
   DataTableSortDirection,
 } from './components/DataTable';
-
-// TODO(next-major): Remove the below types at the next major release
-export type { PageNotificationProps as PageNotificationV2Props } from './components/PageNotification';
-export type { AppNotificationProps as AppNotificationV2Props } from './components/AppNotification';
 
 /**
  * Utilities re-exported from dependent libraries
